+++ Soul Engine/Source Files/Ray Engine/RayEngine.h
@@ -10,7 +10,7 @@ namespace RayEngine {
 
 	//Adds a job to be executed after all updates have taken place. It will execute the 
 	//given function to initialize all its rays and returns the tag that can be used to extract the data in 'UpdateLate'
-	RayJob* AddRayJob(rayType, uint, uint, Camera*,uint);
+	RayJob* AddRayJob(rayType, uint, float, Camera*,uint);
 
 	//A varient that does not copy the results to the CPU but instead returns a cuda* that can be procesed further.
 	//adds a job with a hint to keep its allocated data for ray storage. Speed gains if large ray bundles are given.
@@ -18,7 +18,7 @@ namespace RayEngine {
 
 	//amount cant be changed as it effects storage, but all else can;
 	//return false if it wasn't changed
-	bool ChangeJob(RayJob*, uint, uint, Camera*);
+	bool ChangeJob(RayJob*, uint, float, Camera*);
 
 	bool SwapResults(RayJob*, uint, uint);
